+++ hydromt_wflow/workflows/connect.py
@@ -9,6 +9,8 @@ import pandas as pd
 import xarray as xr
 from shapely.geometry import LineString, MultiLineString, Point
 
+from hydromt_wflow.utils import planar_operation_in_utm
+
 logger = logging.getLogger(__name__)
 
 
@@ -265,7 +267,9 @@ def wflow_1dmodel_connection(
             if not include_river_boundaries:
                 gdf_tributaries = gpd.GeoDataFrame()
         else:
-            gdf_trib["geometry"] = gdf_trib.centroid
+            gdf_trib = gdf_trib.to_crs(ds_model.raster.crs)
+            centroid = planar_operation_in_utm(gdf_trib, lambda geom: geom.centroid)
+            gdf_trib["geometry"] = centroid
 
             # Merge with gdf_tributary if include_river_boundaries
             # else only keep intersecting tributaries